+++ com.listonos.ui/CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added Listonos's Audio System as an optional dependency
 - Added ButtonWithAudio prefab
 - Added AspectRatioFitterLayout behavior
+- Added ButttonWithIcon prefab
+- Added unit tests for navigation system
+- Changed NavigationSystem to require struct instead of Enum to support, for example, int generic param
 - Fixed UI spritesheet visual artifacts
 ## 0.0.0
 - First version of the package